# web/doc/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "University Supervisor API",
        "version": "0.7"
    },
    "paths": {
        "/site/login": {
            "post": {
                "operationId": "0e6ee86b2507b85475c45d4fe785210a",
                "responses": {
                    "200": {
                        "description": "Login"
                    }
                }
            }
        },
        "/admin/course/get-all": {
            "get": {
                "operationId": "5af467b333f33aa849b4d55b928cd1cc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "name": "per-page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список курсов и семестров"
                    }
                }
            }
        },
        "/admin/discipline-time/get-all": {
            "get": {
                "operationId": "381243101bd03618f3faf3ddc9c85644",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "name": "per-page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все часы по всем дисциплинам"
                    }
                }
            }
        },
        "/admin/discipline-time/set-all": {
            "patch": {
                "operationId": "7af1591fbd9563864d012c1ceae64b11",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": [
                                    {
                                        "disciplineId": 1,
                                        "semesterId": 1,
                                        "hours": 99.9
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Изменение часов дисциплин"
                    }
                }
            }
        },
        "/admin/teacher-preference/get-all": {
            "get": {
                "operationId": "fe94d0a95d1d6a784b330c64850ec509",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "name": "per-page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список предпочтений преподавателей"
                    }
                }
            }
        },
        "/admin/teacher-preference/set-all": {
            "patch": {
                "operationId": "1f0733ee5adee921d6879710c5008d61",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": [
                                    {
                                        "teacherId": 1,
                                        "disciplineId": 1,
                                        "semesterId": 1,
                                        "importanceCoefficient": 99.9
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Изменение предпочтений преподавателей"
                    }
                }
            }
        },
        "/admin/teacher-rate/get-all": {
            "get": {
                "operationId": "ab06fc548a6b65961cf0b1318a53a1fc",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "name": "per-page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Все ставки преподавателей"
                    }
                }
            }
        },
        "/admin/teacher-rate/set-all": {
            "patch": {
                "operationId": "1d5ed30dc28fbb6b75598e08da582a06",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": [
                                    {
                                        "teacherId": 1,
                                        "hours": 99.9
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Изменение ставок преподавателей"
                    }
                }
            }
        },
        "/admin/teacher-time-management/get-all": {
            "get": {
                "operationId": "2c0497c2ad75d86dfad0af257b100547",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "name": "per-page",
                        "in": "query"
                    },
                    {
                        "name": "disciplineName",
                        "in": "query"
                    },
                    {
                        "name": "teacherName",
                        "in": "query"
                    },
                    {
                        "name": "semesterName",
                        "in": "query"
                    },
                    {
                        "name": "disciplineId",
                        "in": "query"
                    },
                    {
                        "name": "teacherId",
                        "in": "query"
                    },
                    {
                        "name": "semesterId",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "11Список отданных преподавателям часов по дисциплинам и семестрам"
                    }
                }
            }
        },
        "/admin/teacher-time-management/set-all": {
            "patch": {
                "operationId": "9f9aa6fe8c65ac3e3bc1843a405f43e1",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "example": [
                                    {
                                        "teacherId": 1,
                                        "disciplineId": 1,
                                        "semesterId": 1,
                                        "hours": 99.9
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Изменение часов преподавателей по дисциплинам и семестрам"
                    }
                }
            }
        },
        "/admin/teacher-time-management/generate-new": {
            "put": {
                "operationId": "f273d3b4b597e50bcc03723770705488",
                "responses": {
                    "200": {
                        "description": "Авто-генерация новых часов преподавателей по дисциплинам и семестрам"
                    }
                }
            }
        },
        "/document-builder/site/educational-work": {
            "post": {
                "operationId": "5dab04af5a072a1f8ff2a01e9ec8f7c7",
                "responses": {
                    "200": {
                        "description": "Документ по образовательной работе"
                    }
                }
            }
        }
    }
}